// web/src/app/dataaccess/api/models/V1CreateAccountResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * go_load/v1/go_load.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface V1CreateAccountResponse
 */
export interface V1CreateAccountResponse {
    /**
     * 
     * @type {string}
     * @memberof V1CreateAccountResponse
     */
    accountId?: string;
}

/**
 * Check if a given object implements the V1CreateAccountResponse interface.
 */
export function instanceOfV1CreateAccountResponse(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function V1CreateAccountResponseFromJSON(json: any): V1CreateAccountResponse {
    return V1CreateAccountResponseFromJSONTyped(json, false);
}

export function V1CreateAccountResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): V1CreateAccountResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'accountId': !exists(json, 'accountId') ? undefined : json['accountId'],
    };
}

export function V1CreateAccountResponseToJSON(value?: V1CreateAccountResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'accountId': value.accountId,
    };
}