// web/src/app/dataaccess/api/models/V1CreateAccountRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * go_load/v1/go_load.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface V1CreateAccountRequest
 */
export interface V1CreateAccountRequest {
    /**
     * 
     * @type {string}
     * @memberof V1CreateAccountRequest
     */
    accountName?: string;
    /**
     * 
     * @type {string}
     * @memberof V1CreateAccountRequest
     */
    password?: string;
}

/**
 * Check if a given object implements the V1CreateAccountRequest interface.
 */
export function instanceOfV1CreateAccountRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function V1CreateAccountRequestFromJSON(json: any): V1CreateAccountRequest {
    return V1CreateAccountRequestFromJSONTyped(json, false);
}

export function V1CreateAccountRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): V1CreateAccountRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'accountName': !exists(json, 'accountName') ? undefined : json['accountName'],
        'password': !exists(json, 'password') ? undefined : json['password'],
    };
}

export function V1CreateAccountRequestToJSON(value?: V1CreateAccountRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'accountName': value.accountName,
        'password': value.password,
    };
}